Refactor to support credential unit testing.
